// src/types/receipt.ts
export interface ReceiptModel {
    booking_id: string;
    inv_no: string;
    property_title: string;
    property_address: string;
    guest_full_name: string;
    email: string;
    monthly_fee: number;
    total_paid: number;
    service_fee: number;
    invoice_date: string;
    booking_paid_date: string;
  }
  